add file extensions rich presence + formatting
file extension icon now shows as the large image and large image hover shows as 'Edting a ${fileExt} file'.
aliases are set for edge cases like hpp, h and such which you can preview in the commit.

there are about 25 supported languages and about 30 including edge cases, however, if you think a language should be included and it isnt, please make an issue with the language name and an icon if you would like to suggest one, thanks!

also, if your lannguage is not one that is supported at the moment, the large image will be set to a default file image, in which case you can make an issue. but the large image hover will still say what kind of file you are editing.

# extra stuff
- fix a indentation error on line 12
- edit readme for new features + formatting
- gitignore the icons folder which contains svgs and pngs